fold module graph types directly into Spago.Purs
